wagtail-images/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
anyascii==0.3.2
    # via wagtail
appengine-python-standard==1.1.6
    # via wagtail-images (pyproject.toml)
asgiref==3.8.1
    # via django
attrs==23.2.0
    # via appengine-python-standard
beautifulsoup4==4.12.3
    # via wagtail
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
defusedxml==0.7.1
    # via willow
django==4.2.13
    # via
    #   django-filter
    #   django-modelcluster
    #   django-permissionedforms
    #   django-storages
    #   django-taggit
    #   django-treebeard
    #   djangorestframework
    #   laces
    #   wagtail
    #   wagtail-images (pyproject.toml)
django-filter==23.5
    # via wagtail
django-modelcluster==6.3
    # via wagtail
django-permissionedforms==0.1
    # via wagtail
django-storages[google]==1.14.3
    # via wagtail-images (pyproject.toml)
django-taggit==5.0.1
    # via wagtail
django-treebeard==4.7.1
    # via wagtail
djangorestframework==3.15.1
    # via wagtail
draftjs-exporter==5.0.0
    # via wagtail
et-xmlfile==1.1.0
    # via openpyxl
filetype==1.2.0
    # via willow
frozendict==2.4.4
    # via appengine-python-standard
google-api-core[grpc]==2.19.0
    # via
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-storage
google-auth==2.29.0
    # via
    #   appengine-python-standard
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-core
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-appengine-logging==1.4.3
    # via google-cloud-logging
google-cloud-audit-log==0.2.5
    # via google-cloud-logging
google-cloud-core==2.4.1
    # via
    #   google-cloud-logging
    #   google-cloud-storage
google-cloud-logging==3.10.0
    # via wagtail-images (pyproject.toml)
google-cloud-storage==2.16.0
    # via django-storages
google-crc32c==1.5.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.0
    # via google-cloud-storage
googleapis-common-protos[grpc]==1.63.0
    # via
    #   google-api-core
    #   google-cloud-audit-log
    #   grpc-google-iam-v1
    #   grpcio-status
grpc-google-iam-v1==0.13.0
    # via google-cloud-logging
grpcio==1.63.0
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
grpcio-status==1.62.2
    # via google-api-core
html5lib==1.1
    # via wagtail
idna==3.7
    # via requests
jinja2==3.1.4
    # via wagtail-images (pyproject.toml)
l18n==2021.3
    # via wagtail
laces==0.1.1
    # via wagtail
markupsafe==2.1.5
    # via jinja2
mock==5.1.0
    # via appengine-python-standard
openpyxl==3.1.2
    # via wagtail
pillow==10.3.0
    # via
    #   appengine-python-standard
    #   pillow-heif
    #   wagtail
pillow-heif==0.16.0
    # via willow
proto-plus==1.23.0
    # via
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-logging
protobuf==4.25.3
    # via
    #   appengine-python-standard
    #   google-api-core
    #   google-cloud-appengine-logging
    #   google-cloud-audit-log
    #   google-cloud-logging
    #   googleapis-common-protos
    #   grpc-google-iam-v1
    #   grpcio-status
    #   proto-plus
psycopg[binary]==3.1.19
    # via wagtail-images (pyproject.toml)
psycopg-binary==3.1.19
    # via psycopg
pyasn1==0.6.0
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.0
    # via google-auth
pytz==2024.1
    # via
    #   appengine-python-standard
    #   django-modelcluster
    #   l18n
requests==2.31.0
    # via
    #   appengine-python-standard
    #   google-api-core
    #   google-cloud-storage
    #   wagtail
rsa==4.9
    # via google-auth
ruamel-yaml==0.18.6
    # via appengine-python-standard
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
six==1.16.0
    # via
    #   appengine-python-standard
    #   html5lib
    #   l18n
soupsieve==2.5
    # via beautifulsoup4
sqlparse==0.5.0
    # via django
telepath==0.3.1
    # via wagtail
typing-extensions==4.11.0
    # via psycopg
urllib3==1.26.18
    # via
    #   appengine-python-standard
    #   requests
wagtail==6.0.3
    # via wagtail-images (pyproject.toml)
webencodings==0.5.1
    # via html5lib
willow[heif]==1.8.0
    # via
    #   wagtail
    #   willow
